@@ -24,8 +24,7 @@
                 </div>
             </div>
             <div class="column is-6">
-                <div class="block"></div>
-                <div v-for="pos in parrillaDerecha">
+                <div v-for="pos in parrillaDerecha" class="parrillaDerecha">
                     <StartGridCard :gridPos="pos"/>
                     <div class="block"></div>
                 </div>
@@ -38,12 +37,12 @@
 import {StartGridPosition} from "@/types/StartGridPosition";
 import StartGridCard from "@/components/gps/StartGridCard.vue";
 import {RaceSession} from "@/types/RaceSession";
-import EventBus from "@/plugins/eventbus";
-import dayjs from "@/plugins/dayjs";
 
 import {defineComponent, PropType} from "vue";
-import {useAuthStore} from "@/pinia/authStore";
-import {useCommunityStore} from "@/pinia/communityStore";
+import {useAuthStore} from "@/store/authStore";
+import {useCommunityStore} from "@/store/communityStore";
+import useEmitter from "@/composables/useEmitter";
+import dayjs from "dayjs";
 
 export default defineComponent({
     name: "StartGridList",
@@ -55,12 +54,13 @@ export default defineComponent({
         }
     },
     setup() {
+        const emitter = useEmitter();
         const authStore = useAuthStore();
         const communityStore = useCommunityStore();
 
         const currentUser = authStore.user;
         const currentCommunity = communityStore.community;
-        return { currentUser, currentCommunity };
+        return { currentUser, currentCommunity, emitter };
     },
     data() {
         return {
@@ -80,7 +80,7 @@ export default defineComponent({
     },
     methods: {
         changeGridSession() {
-            EventBus.$emit('changeGridSession', this.chosenSession);
+            this.emitter.emit('changeGridSession', this.chosenSession);
         },
     },
     computed: {
@@ -97,3 +97,9 @@ export default defineComponent({
     }
 });
 </script>
+
+<style scoped>
+.parrillaDerecha {
+    margin-top: 1.5rem;
+}
+</style>
